# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jwt/quotations/create": {
            "post": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Crear una cotización",
                "description": "Este endpoint crea una nueva cotización para un auto, incluyendo los servicios solicitados y su estado de aprobación.",
                "operationId": "022894ef495ded902bab5f14b82a94b5",
                "requestBody": {
                    "description": "Datos para crear una cotización",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "carId",
                                    "services"
                                ],
                                "properties": {
                                    "carId": {
                                        "description": "ID del auto al cual se le crea la cotización",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mechanicId": {
                                        "description": "ID del mecanico que hara la mantencion en caso de aprobar la mantencion",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "services": {
                                        "description": "Lista de servicios solicitados",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "serviceId": {
                                                    "description": "ID del servicio",
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "isApproved": {
                                                    "description": "Si el servicio es aprobado por el cliente",
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "status": {
                                        "description": "Estado de la cotización (aprobada o no)",
                                        "type": "integer",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cotización creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación o de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jwt/quotations/{quotationId}": {
            "get": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Obtiene detalles de una cotización",
                "description": "Devuelve la información de una cotización específica, incluyendo el auto, los servicios aprobados por el cliente y los servicios no aprobados.",
                "operationId": "4174e5d189b35eedde57c0e4f2bc93dd",
                "parameters": [
                    {
                        "name": "quotationId",
                        "in": "path",
                        "description": "ID de la cotización",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cotización obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "car": {
                                            "description": "Detalles del auto",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "make": {
                                                    "type": "string",
                                                    "example": "Toyota"
                                                },
                                                "model": {
                                                    "type": "string",
                                                    "example": "Corolla"
                                                },
                                                "year": {
                                                    "type": "integer",
                                                    "example": 2020
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "quotation": {
                                            "description": "Detalles de la cotización",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.75
                                                },
                                                "car_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "servicesApprovedByClient": {
                                            "description": "Servicios aprobados por el cliente",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Cambio de aceite"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Cambio de aceite del motor"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "servicesNotApprovedByClient": {
                                            "description": "Servicios no aprobados por el cliente",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 11
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Cambio de neumáticos"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Cambio de los neumáticos delanteros"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1500.75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "not user found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jwt/quotations/": {
            "get": {
                "tags": [
                    "Quotations"
                ],
                "summary": "Obtener todas las cotizaciones de los vehículos del usuario autenticado",
                "description": "Este endpoint recupera todas las cotizaciones asociadas con los vehículos de propiedad del usuario autenticado. Devuelve un array de cotizaciones para cada vehículo.",
                "operationId": "68ecad39957c860b05160d8d02ca98c9",
                "responses": {
                    "200": {
                        "description": "Cotizaciones recuperadas con éxito."
                    },
                    "401": {
                        "description": "No autorizado. Usuario no autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un error al recuperar las cotizaciones."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}